% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{class}
\name{IncrementalReadOnlySettings-class}
\alias{IncrementalReadOnlySettings-class}
\title{IncrementalReadOnlySettings class}
\value{
An object of class `IncrementalReadOnlySettings`.
}
\description{
IncrementalReadOnlySettings class
}
\examples{
IncrementalReadOnlySettings(
  a = 1,
  b = "hello",
  c = TRUE,
  d = 5L
)
}
