% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrementalReadOnlySettings-class.R
\name{IncrementalReadOnlySettings}
\alias{IncrementalReadOnlySettings}
\title{Constructor for class IncrementalReadOnlySettings}
\usage{
IncrementalReadOnlySettings(...)
}
\arguments{
\item{...}{Named values to set.}
}
\value{
An object of class \code{\linkS4class{IncrementalReadOnlySettings}}.
}
\description{
Constructor for class IncrementalReadOnlySettings
}
\examples{
IncrementalReadOnlySettings(
  a = 1,
  b = "hello",
  c = TRUE,
  d = 5L
)
}
