% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrementalReadOnlySettings-class.R
\docType{class}
\name{IncrementalReadOnlySettings-class}
\alias{IncrementalReadOnlySettings-class}
\title{IncrementalReadOnlySettings class}
\description{
Refer to the constructor function \code{\link[=IncrementalReadOnlySettings]{IncrementalReadOnlySettings()}}.
}
